Split up `SelectedCubeManager` and `ObjectClickedHook` into react and non-react parts
This fixes issues when running a headless studio
