+++ src/views/animator/Animator.tsx
@@ -5,7 +5,7 @@ import { useKeyComboPressed } from "../../contexts/OptionsContext"
 import { useStudio } from "../../contexts/StudioContext"
 import DcaAnimation from "../../studio/formats/animations/DcaAnimation"
 import { useListenableObject, useListenableObjectNullable } from "../../studio/listenableobject/ListenableObject"
-import { useObjectUnderMouse } from "../../studio/util/ObjectClickedHook"
+import { useObjectUnderMouse } from "../../studio/selections/ObjectClickedHook"
 import AnimatorGumballPropertiesBar from "./components/AnimatorGumballPropertiesBar"
 import AnimatorProperties from "./components/AnimatorProperties"
 import AnimatorScrubBar from "./components/AnimatorScrubBar"
