+++ src/views/modeler/logic/CubePointTracker.ts
@@ -1,11 +1,10 @@
 import { useEffect } from 'react';
 import { BoxBufferGeometry, Camera, EventDispatcher, Group, Mesh, MeshLambertMaterial, Object3D, Vector3 } from "three";
-import { setIntersectType } from '../../../studio/util/ObjectClickedHook';
-import SelectedCubeManager from "../../../studio/util/SelectedCubeManager";
+import { setIntersectThrough, setIntersectType } from '../../../studio/selections/ObjectClickedDataHandler';
+import SelectedCubeManager from "../../../studio/selections/SelectedCubeManager";
 import { scaleMeshToCamera } from '../../../studio/util/Utils';
 import { useStudio } from './../../../contexts/StudioContext';
 import { DCMCube, DCMModel } from './../../../studio/formats/model/DcmModel';
-import { setIntersectThrogh as setIntersectThrough } from './../../../studio/util/ObjectClickedHook';
 
 const _defaultNormalColor = 0x23284d
 const _defaultHighlightColor = 0x15c1d4
