+++ src/views/showcase/Showcase.tsx
@@ -3,9 +3,9 @@ import InfoBar from "../../components/InfoBar"
 import StudioCanvas from "../../components/StudioCanvas"
 import { useStudio } from "../../contexts/StudioContext"
 import { useListenableObject } from "../../studio/listenableobject/ListenableObject"
+import { useObjectUnderMouse } from "../../studio/selections/ObjectClickedHook"
 import { ShowcaseLight } from "../../studio/showcase/ShowcaseLight"
 import ShowcaseView from "../../studio/showcase/ShowcaseView"
-import { useObjectUnderMouse } from "../../studio/util/ObjectClickedHook"
 import AnimatorGumballPropertiesBar from "../animator/components/AnimatorGumballPropertiesBar"
 import ShowcaseEffects from "./components/ShowcaseEffects"
 import ShowcaseSidebar from "./components/ShowcaseSidebar"
